Add files via upload
Bitcoin is a decentralized network and digital currency that uses a peer-to-peer system to verify and process transactions. In this project, I will introduce you to a machine learning model which will help in Bitcoin price prediction using python.
Recently Bitcoin has received a lot of attention from the media and the public due to its recent price hike and sudden fall.  As Bitcoin has been viewed as a financial asset and is traded through many cryptocurrency exchanges like as a stock market, Many researchers have studied various factors that affect the price of Bitcoin and the patterns behind its fluctuations using various analytical and predictive methods. 
Several predictive methods have been studied and compared for the task of Bitcoin price prediction using machine learning. In this project, I will be using the Facebook Prophet Model for the task of Bitcoin price prediction using machine learning by using python. 
The Facebook Prophet Library is an open-source additive regression model made available by Facebook for time-series predictions. While there is a more advanced version of the Prophet called as NeuralProphet which is based on Neural Networks, I used a simplified version only which uses machine learning techniques for the Bitcoin price prediction task. But in near future, I would like to get my hands on Neural Prophet and build a model based on the same.
The data is collected from "Yahoo Finance" and is a historical one.